+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2016-09-10 - 0.1.4
+------------------
+* [new] Add Footer option - *Implemented*.
+
+2016-09-09 - 0.1.3
+------------------
+* [new] Add Events - *Implemented*.
+
 2016-09-08 - 0.1.2
 ------------------
 * [new] Add JS component - *Implemented*.